+++ src/app/questionnaire/page.tsx
@@ -1,0 +1,7 @@
+export default function Questionnaire() {
+    return (
+        <div className='flex flex-col items-center justify-center h-screen space-y-4 p-4 sm:p-8'>
+        <h1 className='text-white'>Questionnaire</h1>
+        </div>
+    )
+}
